Spread utility accross multiple files
Creating new files quickly create circular dependencies with
_excel_to_dict.py.

Spreading these functions accross multiple files eases the dependency
problems.
